[feat] 파일 정리
