+++ SMASHING/Presentation/Profile/UserProfileView.swift
@@ -1,0 +1,8 @@
+//
+//  UserProfileView.swift
+//  SMASHING
+//
+//  Created by 이승준 on 1/15/26.
+//
+
+import Foundation
